+++ src/Dynamic/Dynamic1.svelte
@@ -1,0 +1,7 @@
+<h1>Dynamic route 1</h1>
+<p>
+	Lorem ipsum dolor sit amet, consectetur adipiscing elit. Etiam eget ligula eu lectus lobortis
+	condimentum. Aliquam nonummy auctor massa. Pellentesque habitant morbi tristique senectus et netus
+	et malesuada fames ac turpis egestas. Nulla at risus. Quisque purus magna, auctor et, sagittis ac,
+	posuere eu, lectus. Nam mattis, felis ut adipiscing.
+</p>