+++ CHANGELOG.md
@@ -1,6 +1,16 @@
+<!-- markdownlint-disable MD024 -->
+
 # Change Log
 
 All notable changes to [Elegua](http://github.com/howesteve/elegua) will be documented in this file.
+
+## [v2.20] - 2023-08-15
+
+### Added
+
+- dynamic() method: allows dynamic route rendering
+- Added the preventChange() method that will prevent the user from exiting a route if a condition is met.
+- Added the use:preventUnload() action method: prevents user from closing the window if a condition is met.
 
 ## [v2.10] - 2023-07-22